Добавлен Smart класс
